Work to current point. Includes:
Sections Pager - with three named tabs.

ActiveRacerFragment - A fragment and associated XML, with the screen split between two additional fragments:

ActiveRacerActionFragment - A fragment with buttons and displays, designed for user input about selected racers.

ActiveRacerDisplayDragment - A fragment designed to hold a view allowing for the selection of racers.

TimeButton - A button element that counts the current time, or can be set to a specific time upon click.

Signed-off-by: raj24 <raj24@aber.ac.uk>