Noted a bug in TODO statements Added TODO statements for missing java doc
Changed the way in which grabCheckpoints() and getCheckpoints() work; Basically, grabCheckpoints() is how i'm handling getting data to Fragments. This is far from ideal, as it makes error checking harder and the code less re-usable, but does make it simpler for the purpose it's being used. There's no real good way to talk between fragments, so this is my workaround.

Signed-off-by: raj24 <raj24@aber.ac.uk>